Include unused sidecar HED strings in event-level validation
Previously, sidecar HED strings that are not referenced in the TSV
file were only validated at the string level, not at the event or
dataset levels. This would cause issues with definitions once support
for those is implemented in hed-validator, since definitions will be
included in sidecar HED strings that are not referenced in the TSV.
This commit keeps a log of which sidecar strings have been referenced
in the TSV and dumps the unused ones into the TSV before passing to
the dataset-level validation (which includes event-level).

The tests do not currently pass, since they include invalid strings,
previously skipped, that are now being validated (and failed). I need
to come up with a good solution for this problem.