Move placeholder value validation to hed-validator
Version 3.4.0 of hed-validator provides proper support for placeholder
validation, so we no longer need to include it here.

In order to properly support placeholder validation, the categorical
and value string-level validation loops have been split, so that only
value strings (which are those that are stored as flat strings in the
JSON sidecars) are allowed to have sidecars.