Add version into deno bids-validator.js + create shim to run it
This way anyone interested to use it could just

    git clone --depth 1 -b deno-build http://github.com/bids-standard/bids-validator
    export PATH="$PWD/bids-validator:$PATH"

and start using it and know which version it is:

    ❯ bids-validator -V
    v1.14.6-dev.0-9-gdc61d0f3.deno

An alternative for shim would be just to create it in that branch and then
remove in CI only the .js files.  I would be happy to redo if feels more sensible.